pyproject.toml
[tool.poetry]
name = "rag-code"
version = "0.1.0"
description = ""
authors = ["rohit0221 <rohit0221@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
jedi = "^0.19.2"
networkx = "^3.4.2"
tqdm = "^4.67.0"
py2neo = "^2021.2.4"
python-dotenv = "^1.0.1"
chardet = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
